+++ lib/utils.js
@@ -53,7 +53,7 @@ module.exports = {
 		return req.headers['x-forwarded-for'] || req.connection.remoteAddress;
 	},
 
-	slug_match: (slug) => {
+	url_slug_match: (slug) => {
 		const slug_regex = /^[a-z][a-z0-9_-]+$/i;
 		return slug.match(slug_regex) ? true : false;
 	},
